+++ CHANGELOG.md
@@ -79,6 +79,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Fixed a bug that broke registration of the bus in development env
 
+- Logger modifications
+
 ### Removed
 
 - Removed seeders